delete worker - confirm window

+++ src/main/webapp/worker/list.jsp
@@ -41,7 +41,7 @@
 						<td>${worker.lastName}</td>
 						<td>${worker.salary}</td>
 						<td><a href="${link}">Edytuj</a></td>
-						<td><a href="${deleteLink}">Usuń</a></td>
+						<td><a onclick="return confirm('Czy na pewno usunąć pracownika?')" href="${deleteLink}">Usuń</a></td>
 					</tr>
 				</c:forEach>
 			</tbody>
